added file via upload
this chrome extension is very useful for those people who are working for companies for e commerce or some business those who wants to track the leads on websites , this extension is good for those people who are looking for people online might be interested in their product , so that they browse through the bunch of pages and looking for a potential customer and save them to their system. 
The way it works is that you visit a website for example a  profile on linkedin , and then you open up the extensions, click on the leads tracker and there you can see you get the ability to save input that is any url that you pass into and then hit tab then the tab is saved to the extension like that you can grab the leads and do whatever with it.